+++ public/admin/_app/immutable/nodes/14.C7ibv6lK.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/Cvf9p1Tq2.js";export{e as component};