+++ public/admin/_app/immutable/nodes/17.CH1u8kc1.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/BviCU_Xm2.js";export{e as component};