+++ public/admin/_app/immutable/nodes/18.BUYcg_-y.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/B-5iilaM2.js";export{e as component};