+++ public/admin/_app/immutable/nodes/20.CHgVyNsT.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/DELx213Y.js";export{e as component};