+++ public/admin/_app/immutable/nodes/21.gzJDRFzJ.js
@@ -1,1 +1,0 @@
-import{Vt as e}from"../chunks/CpTjFNj2.js";import{c as t}from"../chunks/DhNL3ZRB.js";import{t as n}from"../chunks/P62kgbz-.js";var r=e({load:()=>i}),i=()=>{throw n(308,`${t}/engines#claude-client`)};export{r as universal};