+++ public/admin/_app/immutable/nodes/24.eMqCJc5L.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/CnYwA8sq.js";export{e as component};