+++ public/admin/_app/immutable/nodes/53.CBVB3MrB.js
@@ -1,1 +1,0 @@
-import{Vt as e}from"../chunks/CpTjFNj2.js";import{c as t}from"../chunks/DhNL3ZRB.js";import{t as n}from"../chunks/P62kgbz-.js";import{t as r}from"../chunks/B2DZBQZu.js";var i=e({load:()=>a}),a=()=>{throw n(308,`${t}/agent-portal`)};export{r as component,i as universal};